updated the smooth scroll effect for IOS
