+++ css/main.css
@@ -18,6 +18,7 @@
 }
 
 html {
+    --scroll-behavior: smooth;
     scroll-behavior: smooth;
 }
