+++ index.html
@@ -29,6 +29,8 @@
 
     gtag('config', 'G-QT5LYM28R8');
     </script>
+    <script src="https://unpkg.com/smoothscroll-anchor-polyfill"></script>
+    <script src="https://unpkg.com/smoothscroll-polyfill/dist/smoothscroll.min.js"></script>
 </head>
 
 <body>
@@ -273,6 +275,5 @@
 
         </div>
     </div>
-    <script src="/scripts/application.js"  type="module"></script>
-    <script src="/scripts/sidebar.js"></script>
+    <script type="module" src="/scripts/application.js" ></script>
 </body>
