--- package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "smoothscroll-polyfill": "^0.4.4"
-  }
-}
