+++ scripts/application.js
@@ -1,6 +1,6 @@
-// import smoothscroll from 'smoothscroll-polyfill';
-
-// smoothscroll.polyfill();
-
 import { initScrollingNavs } from './plugins/scrolling.js';
 initScrollingNavs();
+
+import { initActiveNavLinks } from './plugins/sidebar.js';
+initActiveNavLinks();
+
