+++ scripts/plugins/scrolling.js
@@ -32,3 +32,4 @@ class ScrollingNavs {
     });
   }
 }
+
